requirements.txt
# NBA MCP Synthesis System - Python Dependencies
# Python 3.11+ required

# ============================================
# MCP (Model Context Protocol)
# ============================================
mcp>=0.9.0

# ============================================
# AI Model Clients
# ============================================
anthropic>=0.39.0
openai>=1.51.0
google-generativeai>=0.8.3
ollama>=0.3.3

# ============================================
# AWS Services
# ============================================
boto3>=1.35.54
psycopg2-binary>=2.9.10
awscli>=1.35.20

# ============================================
# Core Dependencies
# ============================================
asyncio>=3.4.3
aiohttp>=3.10.10
pydantic>=2.9.2
python-dotenv>=1.0.1
typing-extensions>=4.12.2

# ============================================
# Data Processing
# ============================================
pandas>=2.2.3
numpy>=1.26.4
sqlalchemy>=2.0.36
pyarrow>=18.0.0

# ============================================
# Mathematical & Symbolic Computing
# ============================================
sympy>=1.13.0              # Symbolic mathematics
scipy>=1.13.1              # Scientific computing
matplotlib>=3.9.2          # Plotting (already listed below)
statsmodels>=0.14.0        # Time series analysis, ARIMA, stationarity tests
pmdarima>=2.0.3            # Auto ARIMA model selection

# ============================================
# Utilities
# ============================================
click>=8.1.7           # CLI tools
rich>=13.9.4           # Beautiful terminal output
tqdm>=4.67.0           # Progress bars
httpx>=0.27.2          # HTTP client
tenacity>=9.0.0        # Retry logic

# ============================================
# Machine Learning & Experimentation
# ============================================
scikit-learn>=1.3.0         # ML algorithms, isotonic regression, calibration
xgboost>=2.0.0              # Gradient boosting for ensemble models
mlflow>=2.9.0              # Experiment tracking, model registry

# ============================================
# Advanced Analytics & Econometrics
# ============================================
linearmodels>=5.3          # Panel data econometrics (fixed/random effects)
pydynpd>=0.2.1             # Dynamic panel data GMM (Arellano-Bond, Blundell-Bond)
pymc>=5.0.0                # Bayesian modeling and probabilistic programming
arviz>=0.15.0              # Exploratory analysis of Bayesian models
dowhy>=0.11.0              # Causal inference and causal graphs
networkx>=3.2.0            # Graph algorithms for DAGs
lifelines>=0.28.0          # Survival analysis (Cox, Kaplan-Meier, parametric)
scikit-survival>=0.21.0    # Machine learning survival models

# ============================================
# Data Quality & Validation
# ============================================
great-expectations>=1.2.5

# ============================================
# Web Scraping & Data Sources
# ============================================
beautifulsoup4>=4.12.3
lxml>=5.3.0
requests>=2.32.3
selenium>=4.26.1          # For dynamic content

# ============================================
# Interactive Dashboards
# ============================================
streamlit>=1.40.1
plotly>=5.24.1
altair>=5.4.1

# ============================================
# Notebook Support
# ============================================
jupyter>=1.1.1
ipykernel>=6.29.5
matplotlib>=3.9.2
seaborn>=0.13.2

# ============================================
# Workflow Orchestration
# ============================================
apache-airflow>=2.10.3  # Optional: only if using Airflow

# ============================================
# Documentation & Collaboration
# ============================================
notion-client>=2.2.1     # Notion API
gspread>=6.1.3           # Google Sheets
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1

# ============================================
# Slack Integration (Optional)
# ============================================
slack-sdk>=3.33.3

# ============================================
# SMS/Text Notifications (Optional)
# ============================================
twilio==9.8.5

# ============================================
# Development & Testing
# ============================================
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0
black>=24.10.0
flake8>=7.1.1
mypy>=1.13.0

# ============================================
# Security Scanning
# ============================================
pre-commit>=3.6.0
detect-secrets>=1.4.0
bandit>=1.7.5

# ============================================
# Logging & Monitoring
# ============================================
structlog>=24.4.0
python-json-logger>=2.0.7
